requirements.txt
gdown==4.6.0
numpy==1.24.1
pandas==1.5.2
Pillow==9.4.0
PyYAML==6.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
